Moved to the Helpers utility for SQL operations
- Still some stuff pending to migrate
- Log IP optional

+++ admin/views/logs.php
@@ -2,17 +2,19 @@
 
 global $wpdb;
 
-$table_logs = $wpdb->prefix . "custom_404_pro_logs";
+$helpers = Helpers::singleton();
+$table_logs = $helpers->table_logs;
 
-$action = $_REQUEST["action"];
-
-if($action === "delete") {
-	if(is_array($_REQUEST["path"])) {
-		$sql_delete = "DELETE FROM " . $table_logs . " WHERE id in (" . implode(",", $_REQUEST["path"]) . ")";
-	} else {
-		$sql_delete = "DELETE FROM " . $table_logs . " WHERE id=" . $_REQUEST["path"] . "";
-	}
-	$wpdb->query($sql_delete);
+if(array_key_exists("action", $_REQUEST)) {
+    $action = $_REQUEST["action"];
+    if($action === "delete") {
+        if(is_array($_REQUEST["path"])) {
+            $sql_delete = "DELETE FROM " . $table_logs . " WHERE id in (" . implode(",", $_REQUEST["path"]) . ")";
+        } else {
+            $sql_delete = "DELETE FROM " . $table_logs . " WHERE id=" . $_REQUEST["path"] . "";
+        }
+        $wpdb->query($sql_delete);
+    }
 }
 
 require_once(__DIR__ . '../../LogsClass.php');
@@ -30,7 +32,7 @@ $logs_table->prepare_items();
         <!-- Now we can render the completed list table -->
         <p class="search-box">
 			<label class="screen-reader-text" for="search_id-search-input">Search</label>
-			<input id="search_id-search-input" type="text" name="s" value="<?php echo $_GET["s"]; ?>" autocomplete="off" />
+			<input id="search_id-search-input" type="text" name="s" value="<?php if(array_key_exists('s', $_GET)) { echo $_GET["s"]; } ?>" autocomplete="off" />
 			<input id="search-submit" class="button" type="submit" name="" value="Search" />
 		</p><br /><br />
         <?php $logs_table->display(); ?>
